# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Agora
  description: >
    Media playback system for Raspberry Pi Zero 2 W — REST API for asset management,
    playback control, CMS integration, and status monitoring. A separate player service
    handles GStreamer pipelines, and an optional CMS client connects to Agora CMS for
    centralized scheduling and fleet management.
  version: 0.9.17
  license:
    name: MIT
servers:
  - url: http://{host}:8000/api/v1
    description: Agora device
    variables:
      host:
        default: agora.local
security:
  - apiKey: []
  - sessionCookie: []
...
paths:
...
  /cms/config:
    get:
      tags: [cms]
      summary: Get CMS connection status
      description: >
        Returns the current CMS connection configuration, service status,
        and whether the device has registered with the CMS.
      operationId: getCmsConfig
      responses:
        "200":
          description: CMS configuration and status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CmsConfigResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [cms]
      summary: Configure CMS connection
      description: >
        Set the CMS server address and port. Automatically strips protocol
        prefixes and restarts the CMS client service. The device will
        connect and register on the next attempt.
      operationId: setCmsConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CmsConfigRequest"
      responses:
        "200":
          description: CMS configuration updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  cms_host:
                    type: string
                  cms_port:
                    type: integer
        "400":
          description: Server address is required.
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key passed via header.
    sessionCookie:
      type: apiKey
      in: cookie
      name: agora_session
      description: Signed session cookie (set after web login).
  responses:
    Unauthorized:
      description: Authentication required.
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: Not authenticated
  schemas:
...
    CmsConfigResponse:
      type: object
      properties:
        cms_host:
          type: string
          description: CMS server hostname or IP.
        cms_port:
          type: integer
          description: CMS server port.
          example: 8080
        has_auth_token:
          type: boolean
          description: Whether the device has registered with the CMS.
        service_active:
          type: boolean
          description: Whether the agora-cms-client systemd service is running.
        configured:
          type: boolean
          description: Whether a CMS host has been set.
    CmsConfigRequest:
      type: object
      required: [cms_host]
      properties:
        cms_host:
          type: string
          description: CMS server hostname or IP address. Protocol prefixes are stripped automatically.
        cms_port:
          type: integer
          description: CMS server port.
          default: 8080